+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Please check the [Upgrading Instructions](UPGRADE.md) when upgrading to a newer 
 
 ### Changed
 - The `tests` now use the same environment variable name for the DB driver as the sample app (`YII2_OAUTH2_SERVER_TEST_DB_DRIVER` -> `YII_DB_DRIVER`). (rhertogh)
+- The namespace and inheritance of `Oauth2OidcBearerTokenResponse` have changed to allow the usage of a more generic `Oauth2BearerTokenResponse`. (rhertogh)
 
 ### Deprecated